// EMS Frontend/src/app/model/newrequirement.ts
export interface NewRequirement {
    id: string;
    salesPerson: string;
    deliveryHead: string;
    requirementRecordDate: string;
    customerName: string;
    opportunity: string;
    ltc: string;
    poReceived: string;
    experience: string;
    numberOfRequirements: string;
    perMonthValue: string;
    deliveryGroup: string;
    closed: string;
    location: string;
}